Resolving PR issues
