+++ package.json
@@ -145,12 +145,15 @@
   },
   "dependencies": {
     "@chainsafe/bls": "^0.1.6",
+    "@chainsafe/eth2.0-config": "^0.1.0",
     "@chainsafe/eth2.0-types": "0.1.0",
     "@chainsafe/ssz": "^0.5.1",
+    "@machinomy/types-ethereumjs-units": "^0.0.2",
     "@types/ethereumjs-abi": "^0.6.3",
     "axios": "0.19.0",
     "electron-log": "^3.0.8",
     "ethereumjs-abi": "^0.6.8",
+    "ethereumjs-units": "^0.2.0",
     "node-cmd": "^3.0.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
